Move ecdsa related code to its provider
